Discard match report under certain circumstances
Only valid session ids are expected for regular
(non-race) gametypes. However we should not limit
reporting to servers that allow connecting only
for logged-in players. If all participated players
(that had a playtime) were logged in, a report
is valid. A duel between two logged-in players
on a public server is a good example.
Thus, consider a report valid from the
match start, but discard it once a bot or an
anonymous player gets some playtime.

Squashed: Lots of fixes for an actual integration